refactor: move the AI session rendering out of webviewContent.ts
Extract the AI session surface cluster (session surface view model,
provider selection, session rows, running animations, AI session
context menu) into webviewAiSessionContent.ts and the shared
escapeAttribute helper into webviewHtmlEscape.ts. webviewContent.ts
keeps workspace cards, project groups, and the boot template, and
re-exports getAiSessionsDiv so existing consumers are unaffected.
No behavior change.
