+++ src/webview/webviewHtmlEscape.ts
@@ -1,0 +1,7 @@
+export function escapeAttribute(value: string): string {
+    return (value || '')
+        .replace(/&/g, '&amp;')
+        .replace(/"/g, '&quot;')
+        .replace(/</g, '&lt;')
+        .replace(/>/g, '&gt;');
+}
